# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EcoTrack API",
        "description": "Système de traçabilité intelligente pour la Supply Chain (Mobile & IoT).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8003",
            "description": "Serveur Local"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Inscription d'un nouvel acteur",
                "description": "Deux types d'inscription : \n                    - Admin : crée une entreprise\n                    - Worker : rejoint une entreprise existante via company_id",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "required": [
                                            "name",
                                            "email",
                                            "password",
                                            "role",
                                            "company_name",
                                            "company_type"
                                        ],
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Franck Admin"
                                            },
                                            "email": {
                                                "type": "string",
                                                "format": "email",
                                                "example": "admin@mail.com"
                                            },
                                            "password": {
                                                "type": "string",
                                                "format": "password",
                                                "example": "password123"
                                            },
                                            "role": {
                                                "type": "string",
                                                "example": "admin"
                                            },
                                            "company_name": {
                                                "type": "string",
                                                "example": "Smart Mobility"
                                            },
                                            "company_type": {
                                                "type": "string",
                                                "enum": [
                                                    "producer",
                                                    "logistics",
                                                    "retailer"
                                                ],
                                                "example": "logistics"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "required": [
                                            "name",
                                            "email",
                                            "password",
                                            "role",
                                            "company_id"
                                        ],
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Jean Worker"
                                            },
                                            "email": {
                                                "type": "string",
                                                "format": "email",
                                                "example": "worker@mail.com"
                                            },
                                            "password": {
                                                "type": "string",
                                                "format": "password",
                                                "example": "password123"
                                            },
                                            "role": {
                                                "type": "string",
                                                "example": "worker"
                                            },
                                            "company_id": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            },
                            "examples": {
                                "admin": {
                                    "summary": "Inscription Admin",
                                    "value": {
                                        "name": "Franck Admin",
                                        "email": "admin@mail.com",
                                        "password": "password123",
                                        "role": "admin",
                                        "company_name": "Smart Mobility",
                                        "company_type": "logistics"
                                    }
                                },
                                "worker": {
                                    "summary": "Inscription Worker",
                                    "value": {
                                        "name": "Jean Worker",
                                        "email": "worker@mail.com",
                                        "password": "password123",
                                        "role": "worker",
                                        "company_id": 1
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|asdasdasdasd"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Connexion pour obtenir un Token",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jean@coffee.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie"
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Entrez votre token reçu lors du login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}